refactor(linter/plugins): rename `oxlint/rule-tester` to `oxlint/plugins-dev` (#18793)
Part of #18610.

Our intent is to move `RuleTester` to a separate package `@oxlint/plugins-dev`.

So rename the current entry point which exports `RuleTester` from `oxlint/rule-tester` to `oxlint/plugins-dev`.
