+++ apps/oxlint/tsdown.config.ts
@@ -37,7 +37,7 @@ export default defineConfig([
   // Main build
   {
     ...commonConfig,
-    entry: ["src-js/cli.ts", "src-js/index.ts", "src-js/plugins.ts", "src-js/rule-tester.ts"],
+    entry: ["src-js/cli.ts", "src-js/index.ts", "src-js/plugins.ts", "src-js/plugins-dev.ts"],
     format: "esm",
     external: [
       // External native bindings
