+++ npm/oxlint/package.json
@@ -43,9 +43,9 @@
       "types": "./dist/plugins.d.ts",
       "default": "./dist/plugins.js"
     },
-    "./rule-tester": {
-      "types": "./dist/rule-tester.d.ts",
-      "default": "./dist/rule-tester.js"
+    "./plugins-dev": {
+      "types": "./dist/plugins-dev.d.ts",
+      "default": "./dist/plugins-dev.js"
     }
   },
   "engines": {
